fix(cases): address version API review feedback
