+++ tests/integration/test_case_versions.py
@@ -26,6 +26,7 @@ from tracecat.cases.enums import (
 )
 from tracecat.cases.schemas import CaseCreate, CaseUpdate
 from tracecat.cases.service import CasesService
+from tracecat.cases.versions.diff import compute_case_version_diff
 from tracecat.db.models import Case, CaseEvent, CaseVersion, User, Workspace
 from tracecat.exceptions import TracecatNotFoundError
 from tracecat.pagination import PageParams
@@ -329,9 +330,18 @@ async def test_case_version_history_pagination_filter_and_compare(
     assert baseline.predecessor is None
     assert baseline.diff is None
 
-    comparison = await service.versions.compare_with_predecessor(
-        case_id=case.id,
-        version_id=summary_versions[1].id,
+    with patch(
+        "tracecat.cases.versions.service.asyncio.to_thread",
+        new=AsyncMock(wraps=asyncio.to_thread),
+    ) as mock_to_thread:
+        comparison = await service.versions.compare_with_predecessor(
+            case_id=case.id,
+            version_id=summary_versions[1].id,
+        )
+    mock_to_thread.assert_awaited_once_with(
+        compute_case_version_diff,
+        "Summary v1",
+        "Summary v2",
     )
     assert comparison is not None
     assert comparison.selected.content == "Summary v2"
